@@ -1,13 +1,13 @@
 # https://peps.python.org/pep-0517
 [build-system]
-requires = ["setuptools", "Cython"]
+requires = ["setuptools==71.1.0", "Cython"]
 build-backend = "setuptools.build_meta"
 
 # https://peps.python.org/pep-0621/
 [project]
 name = "ilpy"
 description = "Python wrappers for popular MIP solvers."
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = { text = "MIT" }
 authors = [
     { email = "funkej@janelia.hhmi.org", name = "Jan Funke" },
@@ -25,6 +25,7 @@ dev = [
     "pytest",
     "pytest-cov",
     "cython",
+    "setuptools==71.1.0",
     "numpy",
     "gurobipy",  # used on CI to confirm equality of results
 ]
@@ -48,7 +49,7 @@ readme = { file = ["README.md"] }
 
 # https://beta.ruff.rs/docs
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 src = ["ilpy"]
 
 [tool.ruff.lint]
